+++ packages/expo-blur/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### 💡 Others
 
+- [Android] Started using expo modules gradle plugin.
+
 ## 14.0.1 — 2024-10-22
 
 _This version does not introduce any user-facing changes._